requirements.txt
opencv-python
google-genai
ultralytics
pillow
numpy
python-dotenv
huggingface_hub